vsock-sample: handle client socket disconnect
The python vsock-sample shows how to send and receive
data in a one-shot manner between the instance and the
enclave.

If the client loops instance-side opening and closing
sockets, the server can get a socket disconnected error
and it shall exit, causing the enclave to terminate.

To fix this, catch this exception and resume the server
loop for new connections. This way, the enclave remains
up as the server does not exit anymore.

Signed-off-by: Alexandru Ciobotaru <alcioa@amazon.com>