Improve vsock_sample for Rust
As of dc4bba3d99bc4988a87ce7775cf1bc554537da16, in the case where
parent instance is the server and the enclave is the client, the
`protocol_helpers::recv_loop` function (running within server) errors
with `ECONNRESET` before it can read the data sent by the client. The
client has forcibly closed both send and receive on the connection.

This is because `VsockSocket::drop` called [shutdown](https://docs.rs/nix/0.29.0/nix/sys/socket/fn.shutdown.html)
with [`Shutdown::Both`](https://docs.rs/nix/0.29.0/nix/sys/socket/enum.Shutdown.html#variant.Both)
(and the `VsockSocket` struct was only used in the client code).

To fix this, the `VsockSocket` struct now carries the context of if it
is the write (client) side or read (server) side of the connection,
and `VsockSocket::drop` uses that mode when invoking shutdown.
Now, the server as well as the client functions make use of `VsockSocket`
to model the lifecycle of the connection.